refact: ArchivePost 모델 분리
